email_service/hms/requirements.txt
serverless-offline>=11.0.0
python-dotenv>=0.21.0
